requirements.txt
absl-py==0.15.0
annotated-types==0.6.0
astunparse==1.6.3
beautifulsoup4==4.12.3
bert4keras==0.7.7
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
filelock==3.13.4
flatbuffers==1.12
gast==0.4.0
gdown==5.1.0
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.5.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.34.1
grpcio-status==1.62.1
h5py==3.1.0
httplib2==0.22.0
idna @ file:///croot/idna_1666125576474/work
importlib_metadata==7.1.0
jieba==0.42.1
joblib==1.4.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.6
MarkupSafe==2.1.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
nlpaug==1.1.11
nlpcda==2.5.8
nltk==3.8.1
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==2.2.1
pillow @ file:///croot/pillow_1707233021655/work
proto-plus==1.23.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.13.0
six==1.15.0
soupsieve==2.5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1700840488771/work
Werkzeug==3.0.2
wrapt==1.12.1
zipp==3.18.1
